backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["1TaylanOzturk <1taylanozturk@protonmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.6"
django-cors-headers = "^3.13.0"
djangorestframework = "^3.13.1"
Pillow = "^9.2.0"
python-decouple = "^3.6"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
